Refactor course module structure and add CourseController
This commit refactors the course module by moving the CourseController to a dedicated controllers directory. It also introduces the CourseController with endpoints for creating courses, retrieving courses by level, and checking user answers. Additionally, new TypeScript interfaces for Course, CourseUnit, and CourseExercise are established, along with empty files for Exercise and Question models to support future development.
